fix(refactor): Address review feedback on architecture and config
Fixes 3 critical issues identified in code review:

1. Configuration behavior change (HIGH RISK)
   Problem: enhance_memories_with_query called without batch_size/retries
   Impact: Behavior inconsistent with historical settings
   Fix: Pass DEFAULT_SCHEDULER_RETRIEVER_BATCH_SIZE and RETRIES explicitly

2. Module dependency inversion (ARCHITECTURE REGRESSION)
   Problem: AdvancedSearcher depended on memos.mem_scheduler.utils
   Impact: Violated decoupling goal (memories should not depend on scheduler)
   Fix: Move extract_json_obj/extract_list_items_in_answer to memos.utils
   Changes:
   - src/memos/utils.py: +163 lines (add utility functions)
   - misc_utils.py: -163 lines, +3 lines (re-export for backward compat)
   - Update imports in 6 files

3. Test coverage migration (NOTED)
   Note: _split_batches tests removed but not migrated
   Action: Defer to future test enhancement

Benefits:
- Maintains historical batch/retry behavior
- Restores proper layering (no memories→scheduler dependency)
- Backward compatible (misc_utils re-exports from memos.utils)

Verification:
✓ No scheduler imports in AdvancedSearcher
✓ Config params passed to enhance_memories_with_query
✓ All existing imports work via re-export
