+++ src/memos/mem_scheduler/utils/misc_utils.py
@@ -1,7 +1,4 @@
-import json
 import os
-import re
-import traceback
 
 from collections import defaultdict
 from functools import wraps
@@ -13,6 +10,8 @@ from memos.log import get_logger
 from memos.mem_scheduler.schemas.message_schemas import (
     ScheduleMessageItem,
 )
+# Re-export from memos.utils for backward compatibility
+from memos.utils import extract_json_obj, extract_list_items, extract_list_items_in_answer
 
 
 logger = get_logger(__name__)
@@ -53,169 +52,6 @@ def is_cloud_env() -> bool:
 
 
 def extract_json_obj(text: str):
-    """
-    Safely extracts JSON from LLM response text with robust error handling.
-
-    Args:
-        text: Raw text response from LLM that may contain JSON
-
-    Returns:
-        Parsed JSON data (dict or list)
-
-    Raises:
-        ValueError: If no valid JSON can be extracted
-    """
-    if not text:
-        raise ValueError("Empty input text")
-
-    # Normalize the text
-    text = text.strip()
-
-    # Remove common code block markers
-    patterns_to_remove = ["json```", "```python", "```json", "latex```", "```latex", "```"]
-    for pattern in patterns_to_remove:
-        text = text.replace(pattern, "")
-
-    # Try: direct JSON parse first
-    try:
-        return json.loads(text.strip())
-    except json.JSONDecodeError as e:
-        logger.info(f"Failed to parse JSON from text: {text}. Error: {e!s}", exc_info=True)
-
-    # Fallback 1: Extract JSON using regex
-    json_pattern = r"\{[\s\S]*\}|\[[\s\S]*\]"
-    matches = re.findall(json_pattern, text)
-    if matches:
-        try:
-            return json.loads(matches[0])
-        except json.JSONDecodeError as e:
-            logger.info(f"Failed to parse JSON from text: {text}. Error: {e!s}", exc_info=True)
-
-    # Fallback 2: Handle malformed JSON (common LLM issues)
-    try:
-        # Try adding missing quotes around keys
-        text = re.sub(r"([\{\s,])(\w+)(:)", r'\1"\2"\3', text)
-        return json.loads(text)
-    except json.JSONDecodeError as e:
-        logger.error(f"Failed to parse JSON from text: {text}. Error: {e!s}")
-        logger.error("Full traceback:\n" + traceback.format_exc())
-        raise ValueError(text) from e
-
-
-def extract_list_items(text: str, bullet_prefixes: tuple[str, ...] = ("- ",)) -> list[str]:
-    """
-    Extract bullet list items from LLM output where each item is on a single line
-    starting with a given bullet prefix (default: "- ").
-
-    This function is designed to be robust to common LLM formatting variations,
-    following similar normalization practices as `extract_json_obj`.
-
-    Behavior:
-    - Strips common code-fence markers (```json, ```python, ``` etc.).
-    - Collects all lines that start with any of the provided `bullet_prefixes`.
-    - Tolerates the "• " bullet as a loose fallback.
-    - Unescapes common sequences like "\\n" and "\\t" within items.
-    - If no bullet lines are found, falls back to attempting to parse a JSON array
-      (using `extract_json_obj`) and returns its string elements.
-
-    Args:
-        text: Raw text response from LLM.
-        bullet_prefixes: Tuple of accepted bullet line prefixes.
-
-    Returns:
-        List of extracted items (strings). Returns an empty list if none can be parsed.
-    """
-    if not text:
-        return []
-
-    # Normalize the text similar to extract_json_obj
-    normalized = text.strip()
-    patterns_to_remove = ["json```", "```python", "```json", "latex```", "```latex", "```"]
-    for pattern in patterns_to_remove:
-        normalized = normalized.replace(pattern, "")
-    normalized = normalized.replace("\r\n", "\n")
-
-    lines = normalized.splitlines()
-    items: list[str] = []
-    seen: set[str] = set()
-
-    for raw in lines:
-        line = raw.strip()
-        if not line:
-            continue
-
-        matched = False
-        for prefix in bullet_prefixes:
-            if line.startswith(prefix):
-                content = line[len(prefix) :].strip()
-                content = content.replace("\\n", "\n").replace("\\t", "\t").replace("\\r", "\r")
-                if content and content not in seen:
-                    items.append(content)
-                    seen.add(content)
-                matched = True
-                break
-
-        if matched:
-            continue
-
-    if items:
-        return items
-    else:
-        logger.error(f"Fail to parse {text}")
-
-    return []
-
-
-def extract_list_items_in_answer(
-    text: str, bullet_prefixes: tuple[str, ...] = ("- ",)
-) -> list[str]:
-    """
-    Extract list items specifically from content enclosed within `<answer>...</answer>` tags.
-
-    - When one or more `<answer>...</answer>` blocks are present, concatenates their inner
-      contents with newlines and parses using `extract_list_items`.
-    - When no `<answer>` block is found, falls back to parsing the entire input with
-      `extract_list_items`.
-    - Case-insensitive matching of the `<answer>` tag.
-
-    Args:
-        text: Raw text that may contain `<answer>...</answer>` blocks.
-        bullet_prefixes: Accepted bullet prefixes (default: strictly `"- "`).
-
-    Returns:
-        List of extracted items (strings), or an empty list when nothing is parseable.
-    """
-    if not text:
-        return []
-
-    try:
-        normalized = text.strip().replace("\r\n", "\n")
-        # Ordered, exact-case matching for <answer> blocks: answer -> Answer -> ANSWER
-        tag_variants = ["answer", "Answer", "ANSWER"]
-        matches: list[str] = []
-        for tag in tag_variants:
-            matches = re.findall(rf"<{tag}>([\\s\\S]*?)</{tag}>", normalized)
-            if matches:
-                break
-        # Fallback: case-insensitive matching if none of the exact-case variants matched
-        if not matches:
-            matches = re.findall(r"<answer>([\\s\\S]*?)</answer>", normalized, flags=re.IGNORECASE)
-
-        if matches:
-            combined = "\n".join(m.strip() for m in matches if m is not None)
-            return extract_list_items(combined, bullet_prefixes=bullet_prefixes)
-
-        # Fallback: parse the whole text if tags are absent
-        return extract_list_items(normalized, bullet_prefixes=bullet_prefixes)
-    except Exception as e:
-        logger.info(f"Failed to extract items within <answer> tags: {e!s}", exc_info=True)
-        # Final fallback: attempt direct list extraction
-        try:
-            return extract_list_items(text, bullet_prefixes=bullet_prefixes)
-        except Exception:
-            return []
-
-
 def parse_yaml(yaml_file: str | Path):
     yaml_path = Path(yaml_file)
     if not yaml_path.is_file():
